feat: fix some types and refactor for better separation of concerns
