+++ lib/index.ts
@@ -4,6 +4,6 @@ export * from './mixins/stringify-safe'
 export * from './utils/normalize-select'
 export * from './utils/resolve-circular'
 
-export * from './common-types'
+export * from './types'
 export { ContentfulClientApi } from './create-contentful-api'
 export * from './query'
