+++ lib/types/collection.ts
@@ -1,0 +1,6 @@
+export interface ContentfulCollection<T> {
+  total: number
+  skip: number
+  limit: number
+  items: Array<T>
+}
